requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0  # Para crear APIS REST

# Documentacion de API ( no afecta en nada )
coreapi==2.3.3  
coreschema==0.0.4